+++ preproc/check/check_schism_bdry.m
@@ -1,22 +1,22 @@
-function check_schism_bdry(Mobj, DS, BdryCnd, varName, idx_time, obc_bnds)
+function check_schism_bdry(Mobj, DS, BdryCnd, varName, timeTick, obc_bnds)
 % Check the interpolation at the boundary nodes
 %
 %% Syntax
 % check_schism_bdry(Mobj, DS, BdryCnd, varName)
-% check_schism_bdry(Mobj, DS, BdryCnd, varName, idx_time)
-% check_schism_bdry(Mobj, DS, BdryCnd, varName, idx_time, obc_bnds)
+% check_schism_bdry(Mobj, DS, BdryCnd, varName, timeTick)
+% check_schism_bdry(Mobj, DS, BdryCnd, varName, timeTick,  obc_bnds)
 %
 %% Description
 % check_schism_bdry(Mobj, DS, BdryCnd, varName) checks the interpolation along open boundary nodes
-% check_schism_bdry(Mobj, DS, BdryCnd, varName, idx_time) specifies the time steps
-% check_schism_bdry(Mobj, DS, BdryCnd, varName, idx_time, obc_bnds) specifies the open boundaries
+% check_schism_bdry(Mobj, DS, BdryCnd, varName, timeTick) specifies the time moment
+% check_schism_bdry(Mobj, DS, BdryCnd, varName, timeTick, obc_bnds) specifies the open boundaries
 %
 %% Examples 
 % DS = prep_schism_bdry(Mobj, 'hycom_bys'); 
 % varList = {'ssh', 'temp', 'salt', 'uvel', 'vvel'}; 
 % BdryCnd = interp_schism_bdry(Mobj, DS, varList);
 % 
-% check_schism_bdry(Mobj, DS, BdryCnd, 'temp',1)
+% check_schism_bdry(Mobj, DS, BdryCnd, 'temp', Mobj.time(1))
 %
 %% Input Arguments
 % Mobj - mesh object; datastruct
@@ -25,8 +25,8 @@ function check_schism_bdry(Mobj, DS, BdryCnd, varName, idx_time, obc_bnds)
 %       the datastruct containing boundary data, resulting from "prep_schism_bdry"
 % BdryCnd - data object; datastruct
 %       the datastruct containing boundary data, resulting from "interp_schism_bdry"
-% idx_time - index time (optional); numeric
-%       the index of time steps. Default: idx_time = 1.
+% timeTick - time moment (optional); datetime
+%       the time moment to show the results. Default: timeTick = Mobj.time(1)
 % obc_bnds - open boundary index (optional); numeric
 %       the index of open boundary segments. e.g., obc_bnds =
 %       1:Mobj.obc_counts. All available open boundary segments in "DS"
@@ -37,17 +37,17 @@ function check_schism_bdry(Mobj, DS, BdryCnd, varName, idx_time, obc_bnds)
 %
 %% Author Info
 % Created by Wenfan Wu, Virginia Institute of Marine Science in 2022. 
-% Last Updated on 1 Apr 2025. 
+% Last Updated on 18 Apr 2025. 
 % Email: wwu@vims.edu
 % 
 % See also: check_schism_init
 
 %% Parse inputs
 if nargin < 4; varName = 'temp'; end
-if nargin < 5; idx_time = 1; end
+if nargin < 5; timeTick = Mobj.time(1); end
 
 D = DS.(varName);
-% Determine the used open boundary segments
+% determine the open boundary segments
 obc_bnds_all = find(ismember(Mobj.obc_nodes(1,:), D.ind));
 
 if nargin < 6; obc_bnds = 'all'; end
@@ -69,11 +69,12 @@ nz_raw = numel(D.depth);
 v1 = BdryCnd.(varName);
 v2 = D.var; 
 
-var_new = squeeze(v1(:,:,idx_time))';
-dep_new = -abs(Mobj.depLayers(:, obc_nodes));
+idx_time1 = minfind(Mobj.time, timeTick);
+var_new = squeeze(v1(:,:,idx_time1))';
+dep_new = fillmissing(-abs(Mobj.depLayers(:, obc_nodes)), 'previous',1);
 dist_new = repmat(1:nps, nz_new, 1);
 
-idx_time2 = minfind(D.time, Mobj.time(idx_time));  % time index in the raw data.
+idx_time2 = minfind(D.time, Mobj.time(idx_time1));  % time index in the raw data.
 var_raw = squeeze(v2(:,:, idx_time2))';
 dep_raw = -abs(repmat(D.depth(:), 1, nps));
 dist_raw =  repmat(1:nps, nz_raw, 1);
@@ -83,6 +84,8 @@ figure('Color', 'w')
 subplot(211)
 pcolor(dist_new, dep_new, var_new)
 shading flat
+hold on
+plot(-abs(Mobj.depth(obc_nodes)), 'LineWidth', 1, 'Color', 'k')
 colormap(jet(25))
 colorbar
 cm = caxis; %#ok<*CAXIS>
@@ -90,11 +93,13 @@ ym = ylim;
 xlabel('Along open boundary nodes', 'FontWeight','bold')
 ylabel('Depth (m)', 'FontWeight','bold')
 set(gca, 'Layer', 'top')
-title(['SCHISM (', datestr(Mobj.time(idx_time), 'yyyy-mm-dd'), ')']) %#ok<*DATST>
+title(['SCHISM (', datestr(Mobj.time(idx_time1), 'yyyy-mm-dd HH:MM:SS'), ')']) %#ok<*DATST>
 
 subplot(212)
 pcolor(dist_raw, dep_raw, var_raw)
 shading flat
+hold on
+plot(-abs(Mobj.depth(obc_nodes)), 'LineWidth', 1, 'Color', 'k')
 colormap(jet(25))
 colorbar
 caxis(cm)
@@ -102,7 +107,7 @@ ylim(ym)
 xlabel('Along open boundary nodes', 'FontWeight','bold')
 ylabel('Depth (m)', 'FontWeight','bold')
 set(gca, 'Layer', 'top')
-title(['Raw Data (', datestr(D.time(idx_time2), 'yyyy-mm-dd'), ')'])
+title(['Raw Data (', datestr(D.time(idx_time2), 'yyyy-mm-dd HH:MM:SS'), ')'])
 
 end